apps/recognition/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.9.2
pydantic-settings==2.5.2

# Visão computacional
insightface==0.7.3
onnxruntime==1.19.2
opencv-python-headless==4.10.0.84
numpy==1.26.4
Pillow==10.4.0

# AdaFace backend (opcional — só carrega se FACE_BACKEND=adaface)
# mediapipe traz BlazeFace (Apache 2.0) como detector limpo de licença.
mediapipe==0.10.14
# torch.onnx.export (PyTorch recente) exige onnxscript
onnxscript>=0.2.0

# Vector DB
qdrant-client==1.11.3

# Utilitários
python-dotenv==1.0.1
loguru==0.7.2
